+++ examples/spot/market/getV3LastestKline.js
@@ -2,9 +2,19 @@
 
 const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
 
+
 const bitmartSpotAPI = new BitmartSpotAPI()
 
-bitmartSpotAPI.getTickerDetail('BTC_USDT')
+const before = Math.floor(Date.now() / 1000)
+const after = before - 60 * 60
+
+
+bitmartSpotAPI.getV3LatestKline('BTC_USDT', {
+  before: before,
+  after: after,
+  step: 1,
+  limit: 100,
+})
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))