+++ examples/spot/market/getV3Tickers.js
@@ -6,7 +6,7 @@ const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
 const bitmartSpotAPI = new BitmartSpotAPI()
 
 
-bitmartSpotAPI.getKlineStep()
+bitmartSpotAPI.getV3Tickers()
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))