+++ examples/spot/market/getV3Trades.js
@@ -2,11 +2,12 @@
 
 const BitmartSpotAPI = require('../../../src/bitmartSpotAPI')
 
-
 const bitmartSpotAPI = new BitmartSpotAPI()
 
 
-bitmartSpotAPI.getTicker()
+bitmartSpotAPI.getV3Trades('BTC_USDT', {
+  limit: 5
+})
   .then(response => bitmartSpotAPI.logger.log(response.data))
   .catch(error => bitmartSpotAPI.logger.log(error))